chore: fix clippy warnings in test code (clone_on_copy, field_reassign, needless_range_loop)
QA pass with --tests caught test-only clippy warnings across 40+ files:
- Replace .clone() on Copy types with direct copy
- Use struct init syntax where applicable, allow where deeply nested
- Replace index loops with iterators
- Remove unnecessary casts and duplicate cfg attributes

+++ golden-master/tests/play_note_render_golden.rs
@@ -196,7 +196,7 @@ fn capture_snapshot(
     NoteDrawSnapshot {
         commands: result.commands.iter().map(convert_command).collect(),
         lift_offset_y: result.lift_offset_y,
-        lanecover_offset_y: (result.lanecover_offset_y as f32 * 100.0).round() / 100.0,
+        lanecover_offset_y: (result.lanecover_offset_y * 100.0).round() / 100.0,
     }
 }
